Removed migration & plsql rules
Migration rules are very specific to individual projects so its
better for engines to not have this. Also plsql isn't so often
checked in with java repos anymore and if they are they would
probably like their own rules instead of the defaults which can
be very strict.

So to avoid these, we're just going to get rid of these rules and
keep the ones which more than 90%** of folks would use.

** - just a number out of a hat :)